requirements.txt
mysql-connector-python>=8.0.0
pymongo>=4.0.0
tabulate>=0.8.9
requests>=2.25.0
python-dotenv>=0.19.0 